pyproject.toml
[project]
name = "evalforge"
version = "0.1.0"
description = "Automated eval dataset generator from production LLM logs"
requires-python = ">=3.11"
dependencies = [
  "psycopg[binary]>=3.2",
  "sentence-transformers>=3.0",
  "hdbscan>=0.8.40",
  "umap-learn>=0.5",
  "scikit-learn>=1.5",
  "numpy>=1.26,<2.2",
  "pandas>=2.2",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "rich>=13.7",
  "typer>=0.12",
  "streamlit>=1.36",
  "altair>=5.3",
  "datasets>=3.0",
  "huggingface-hub>=0.25",
  "apscheduler>=3.10",
]

[dependency-groups]
dev = [
  "pytest>=8.2",
  "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/evalforge"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
  "B008",   # typer.Option in defaults is the typer pattern
  "E501",   # long lines OK in SQL/docstrings
  "SIM102", # nested ifs intentional in streamlit (button click + action result)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning:streamlit.*",
  "ignore::UserWarning:umap.*",
]
